backend/requirements.txt
# FastAPI + Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Google Cloud
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.18.0
firebase-admin>=6.4.0

# Gemini API
google-genai>=1.0.0

# Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Security
python-jose[cryptography]>=3.3.0

# Rate Limiting
slowapi>=0.1.9

# Retry Logic
tenacity>=8.2.0
